+++ base/fields/__init__.py
@@ -1,0 +1,10 @@
+# -*- coding: utf-8 -*-
+
+from .base import BaseFileField
+from .base import ChileanRUTField
+
+
+__all__ = [
+    'BaseFileField',
+    'ChileanRUTField',
+]
