+++ base/fields/base.py
@@ -13,28 +13,10 @@ from django.core.exceptions import ValidationError
 
 # utils
 from base import utils
+from .functions import file_path
 
 # translations
 from django.utils.translation import ugettext_lazy as _
-
-
-# public methods
-def file_path(self, name):
-    """
-    Generic method to give to a FileField or ImageField in it's upload_to
-    parameter.
-
-    This returns the name of the class, concatenated with the id of the
-    object and the name of the file.
-    """
-    base_path = "{}/{}/{}/{}"
-
-    return base_path.format(
-        self.__class__.__name__,
-        str(utils.today()),
-        utils.random_string(30),
-        name
-    )
 
 
 class ChileanRUTField(CharField):
